# Reduce LZMA compression for faster therock-archives builds.