Progress on https://github.com/ROCm/TheRock/issues/726.

That issue has more data, but the most relevant experiment was generating `amd-llvm_dev_generic.tar.xz` on my dev machine with different compression settings:

preset | size | time | notes
-- | -- | -- | --
0 | 139.8 MB | 15.6 seconds |
1 | 123.9 MB | 17.6 seconds |
2 | 115.3 MB | 23.0 seconds | Proposal A
3 | 109.5 MB | 35.7 seconds | Proposal B
4 | 107.6 MB | 66.9 seconds |
5 | 98.24 MB | 107.2 seconds |
6 | 95.14 MB | 140.7 seconds | Default
7 | 91.75 MB | 158.7 seconds |
8 | 89.44 MB | 175.7 seconds |
9 | 87.47 MB | 183.7 seconds |

Timing before (https://github.com/ROCm/TheRock/actions/runs/15327687672):

job | `therock-dist` time | `therock-archives` time
-- | --: | --:
Linux gfx94x | 1h 35m | 6m 50s
Linux gfx110x | 1h 50m | 4m 20s
Windows gfx110x | 50m | 1h 1m 30s

Timing after (https://github.com/ROCm/TheRock/actions/runs/15327687672):

job | `therock-dist` time | `therock-archives` time
-- | --: | --:
Linux gfx94x | 1h 51m  | 2m35s
Linux gfx110x | 2h 20m | 1m 13s
Windows gfx110x | 48m | 10m 17s